Share host/participant info displaying.
